+++ airline-core/src/main/moditect/module-info.java
@@ -19,6 +19,7 @@ module com.github.rvesse.airline
   requires com.github.rvesse.airline.io;
   requires org.apache.commons.lang3;
   requires org.apache.commons.collections4;
+  requires io.github.classgraph;
 
   // Optional dependencies that a user might choose to add
   requires static java.inject;